refactor: reorganize orchestrator module
