+++ src/orchestrator/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Orchestrator module exports.
+ */
+
+export { type InspectOptions, inspectFiles, inspectProject } from "./executor.ts";
